+++ CHANGES.rst
@@ -24,6 +24,7 @@ in progress
 - Support reading data from HTTP
 - Performance: Use the Hunter code tracing toolkit to trace execution path
 - Performance: Don't open resource twice when using ``frictionless`` backend
+- Performance: Improve peeking into Parquet files
 
 
 2023-01-xx 0.0.0